Add puppet-bolt@2 Cask for users who aren't ready to upgrade
We're planning to release Bolt 3.0 shortly. With most package managers
users can specify an older Bolt version if they aren't ready to upgrade,
but with homebrew we need to provide a separate puppet-bolt@2 package
for users who aren't ready to join the 3.x stream. This stream should
never be updated, it just exists so that homebrew users can easily
install Bolt 2.x.